Split ymodel.ts into multiple files
