+++ javascript/src/index.ts
@@ -8,5 +8,17 @@
  */
 
 export * from './api.js';
-export * from './ymodels.js';
 export * from './utils.js';
+
+export * from './ytext.js';
+export * from './ydocument.js';
+export * from './yfile.js';
+export * from './ynotebook.js';
+export {
+	YCellType,
+	YBaseCell,
+	YRawCell,
+	YMarkdownCell,
+	YCodeCell,
+	createStandaloneCell
+} from './ycell.js';
